+++ R/utils-unicode.R
@@ -1,10 +1,13 @@
 #' Other Unicode utilities
 #'
 #' `ucp2label()` returns Unicode code point \dQuote{labels} as a character vector.
+#' `ucp_sort()` sorts Unicode code points.
 #' `is_combining_character()` returns `TRUE` if the character is a \dQuote{combining} character.
 #'
 #' @param x A character vector of Unicode code points.
+#' @param decreasing If `TRUE` do a decreasing sort.
 #' @return `ucp2label()` returns a character vector of Unicode labels.
+#'         `ucp_sort()` returns a character vector of Unicode code points.
 #'         `is_combining_character()` returns a logical vector.
 #' @examples
 #'   # Get the Unicode Code Point "label" for "R"
@@ -12,11 +15,18 @@
 #'
 #'   is_combining_character(str2ucp("a"))
 #'   is_combining_character("U+0300") # COMBINING GRAVE ACCENT
-#' @seealso [hex2ucp()], [int2ucp()], [name2ucp()], and [str2ucp()] all return Unicode code points.
+#' @seealso [block2ucp()], [hex2ucp()], [int2ucp()], [name2ucp()], [range2ucp()], and [str2ucp()]
+#'           all return Unicode code points.
 #' @rdname unicode_utilities
 #' @export
 ucp2label <- function(x) {
     Unicode::u_char_label(x)
+}
+
+#' @rdname unicode_utilities
+#' @export
+ucp_sort <- function(x, decreasing = FALSE) {
+    int2ucp(sort(as.integer(Unicode::as.u_char(x)), decreasing = decreasing))
 }
 
 #' @rdname unicode_utilities